Iterating over the module structure
