+++ state/src/raw.rs
@@ -1,4 +1,3 @@
-use crate::interface::*;
 use crate::StateErr;
 use atom::prelude::*;
 use atom::space::*;
@@ -102,6 +101,24 @@ impl<'a> StoreHandle<'a> {
     }
 }
 
+pub struct StatePropertyWriter<'a> {
+    head: SpaceHead<'a>,
+}
+
+impl<'a> StatePropertyWriter<'a> {
+    pub fn new(head: SpaceHead<'a>) -> Self {
+        Self { head }
+    }
+
+    pub fn init<'b, A: Atom<'a, 'b>>(
+        &'b mut self,
+        urid: URID<A>,
+        parameter: A::WriteParameter,
+    ) -> Option<A::WriteHandle> {
+        (&mut self.head as &mut dyn MutSpace).init(urid, parameter)
+    }
+}
+
 pub struct RetrieveHandle<'a> {
     retrieve_fn: sys::LV2_State_Retrieve_Function,
     handle: sys::LV2_State_Handle,
@@ -141,6 +158,29 @@ impl<'a> RetrieveHandle<'a> {
         };
 
         Some(StatePropertyReader::new(type_, Space::from_slice(space)))
+    }
+}
+
+pub struct StatePropertyReader<'a> {
+    type_: URID,
+    body: Space<'a>,
+}
+
+impl<'a> StatePropertyReader<'a> {
+    pub fn new(type_: URID, body: Space<'a>) -> Self {
+        Self { type_, body }
+    }
+
+    pub fn read<A: Atom<'a, 'a>>(
+        &self,
+        urid: URID<A>,
+        parameter: A::ReadParameter,
+    ) -> Option<A::ReadHandle> {
+        if urid == self.type_ {
+            A::read(self.body, parameter)
+        } else {
+            None
+        }
     }
 }
